Move icon mapper and getNotificationUX into a utils file
